backend/requirements.txt
fastapi
uvicorn[standard]
watchdog
pdfplumber
pytesseract
pdf2image
openai
supabase
pydantic
python-dotenv
python-multipart
requests
